pyproject.toml
[project]
name = "uploading-images-llm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.19",
    "langchain-community>=0.3.18",
    "langchain-core>=0.3.37",
    "langchain-google-genai>=2.0.10",
    "langchain-openai>=0.3.6",
    "pdf2image>=1.17.0",
    "pillow>=11.1.0",
    "pymupdf>=1.25.3",
    "pytesseract>=0.3.13",
]
